Haal dubbele code weg dmv ThaliappActivity
